Added Files "Sensors"
Files .h .c added for sensors.

Code compiling without warning or mistakes.
