+++ Core/Inc/user_def_screen.h
@@ -1,0 +1,15 @@
+#ifndef INC_USER_DEF_SCREEN_H_
+#define INC_USER_DEF_SCREEN_H_
+
+/* SCREEN includes -----------------------------------------------------------*/
+#include "main.h"
+
+
+/* SCREEN structures ---------------------------------------------------------*/
+
+
+/* SCREEN functions declaration ----------------------------------------------*/
+//SCREEN GEN
+
+
+#endif /* INC_USER_DEF_SCREEN_H_ */
