+++ Core/Inc/user_def_sdcard.h
@@ -1,0 +1,15 @@
+#ifndef INC_USER_DEF_SDCARD_H_
+#define INC_USER_DEF_SDCARD_H_
+
+/* SDCARD includes -----------------------------------------------------------*/
+#include "main.h"
+
+
+/* SDCARD structures ---------------------------------------------------------*/
+
+
+/* SDCARD functions declaration ----------------------------------------------*/
+//SDCARD GEN
+
+
+#endif /* INC_USER_DEF_SDCARD_H_ */
