+++ Core/Src/user_def_screen.c
@@ -1,0 +1,16 @@
+/* SCREEN includes -----------------------------------------------------------*/
+#include <stdio.h>
+#include <string.h>
+
+#include "user_def_screen.h"
+
+
+/* SCREEN define -------------------------------------------------------------*/
+
+
+/* SCREEN variables ----------------------------------------------------------*/
+
+
+/* SCREEN functions ----------------------------------------------------------*/
+//SCREEN GEN
+
