+++ Core/Src/user_def_sdcard.c
@@ -1,0 +1,16 @@
+/* SDCARD includes -----------------------------------------------------------*/
+#include <stdio.h>
+#include <string.h>
+
+#include "user_def_sdcard.h"
+
+
+/* SDCARD define -------------------------------------------------------------*/
+
+
+/* SDCARD variables ----------------------------------------------------------*/
+
+
+/* SDCARD functions ----------------------------------------------------------*/
+//SDCARD GEN
+
