+++ Core/Src/user_def_sensors.c
@@ -1,0 +1,269 @@
+/* SENSORS includes ----------------------------------------------------------*/
+#include <stdio.h>
+#include <string.h>
+#include "i2c.h"
+#include "tim.h"
+#include "usart.h"
+#include "hts221_reg.h"
+#include "lps22hh_reg.h"
+#include "user_def_sensors.h"
+
+
+/* SENSORS define ------------------------------------------------------------*/
+#define SENSOR_BUS hi2c1
+#define BOOT_TIME 5
+#define TX_BUF_DIM 1000
+#define ADC_BUF_LEN 1
+
+
+/* SENSORS variables ---------------------------------------------------------*/
+static uint8_t whoamI, rst;
+static uint8_t tx_buffer[TX_BUF_DIM];
+static uint16_t adc_buf[ADC_BUF_LEN];
+
+stmdev_ctx_t dev_ctx_hts221;
+stmdev_ctx_t dev_ctx_lps22hh;
+lin_t lin_hum;
+lin_t lin_temp;
+
+static int16_t data_raw_humidity;
+static int16_t data_raw_temperature;
+static uint32_t data_raw_pressure;
+
+static float humidity_perc;
+static float temperature_degC;
+static float pressure_hPa;
+static float windspeed_kph;
+static float rainfall_mm;
+
+
+/* SENSORS functions ---------------------------------------------------------*/
+//SENSORS GEN
+void tx_com(uint8_t *tx_buffer, uint16_t len) {
+	HAL_UART_Transmit(&huart1, tx_buffer, len, 1000);
+}
+
+void platform_delay(uint32_t ms) {
+	HAL_Delay(ms);
+}
+
+void SENSORS_Start_Conversion(void) {
+	// start the temperature conversion
+	hts221_one_shoot_trigger_set(&dev_ctx_hts221, PROPERTY_ENABLE);
+	uint8_t ctrl_reg2 = 0;
+	lps22hh_read_reg(&dev_ctx_lps22hh, LPS22HH_CTRL_REG2, &ctrl_reg2, 1);
+	ctrl_reg2 |= 1;
+	lps22hh_write_reg(&dev_ctx_lps22hh, LPS22HH_CTRL_REG2, &ctrl_reg2, 1);
+}
+
+float linear_interpolation(lin_t *lin, int16_t x) {
+	return ((lin->y1 - lin->y0) * x
+			+ ((lin->x1 * lin->y0) - (lin->x0 * lin->y1))) / (lin->x1 - lin->x0);
+}
+
+//SENSOR hts221
+void SENSOR_hts221_Init(void) {
+	dev_ctx_hts221.write_reg = hts221_write;
+	dev_ctx_hts221.read_reg = hts221_read;
+	dev_ctx_hts221.handle = & SENSOR_BUS;
+	/* Check device ID */
+	whoamI = 0;
+	int result = hts221_device_id_get(&dev_ctx_hts221, &whoamI);
+
+	if (result == 0) {
+		printf("WHO_AM_I: %d\r\n", whoamI);
+	} else {
+		printf("Error reading WHO_AM_I register\r\n");
+	}
+
+	if (whoamI != HTS221_ID)
+		while (1)
+			; /*manage here device not found */
+
+	/* Read humidity calibration coefficient */
+	hts221_hum_adc_point_0_get(&dev_ctx_hts221, &lin_hum.x0);
+	hts221_hum_rh_point_0_get(&dev_ctx_hts221, &lin_hum.y0);
+	hts221_hum_adc_point_1_get(&dev_ctx_hts221, &lin_hum.x1);
+	hts221_hum_rh_point_1_get(&dev_ctx_hts221, &lin_hum.y1);
+
+	/* Read temperature calibration coefficient */
+	hts221_temp_adc_point_0_get(&dev_ctx_hts221, &lin_temp.x0);
+	hts221_temp_deg_point_0_get(&dev_ctx_hts221, &lin_temp.y0);
+	hts221_temp_adc_point_1_get(&dev_ctx_hts221, &lin_temp.x1);
+	hts221_temp_deg_point_1_get(&dev_ctx_hts221, &lin_temp.y1);
+
+	/* Enable Block Data Update */
+	hts221_block_data_update_set(&dev_ctx_hts221, PROPERTY_ENABLE);
+
+	/* Set Output Data Rate */
+	hts221_data_rate_set(&dev_ctx_hts221, HTS221_ONE_SHOT);
+
+	hts221_drdy_on_int_set(&dev_ctx_hts221, PROPERTY_ENABLE);
+
+	/* Device power on */
+	hts221_power_on_set(&dev_ctx_hts221, PROPERTY_ENABLE);
+}
+
+int32_t hts221_write(void *handle, uint8_t reg, const uint8_t *bufp,
+		uint16_t len) {
+	reg |= 0x80;
+	HAL_I2C_Mem_Write(handle, HTS221_I2C_ADDRESS, reg, I2C_MEMADD_SIZE_8BIT,
+			(uint8_t*) bufp, len, 1000);
+	return 0;
+}
+
+int32_t hts221_read(void *handle, uint8_t reg, uint8_t *bufp,
+		uint16_t len) {
+	/* Read multiple command */
+	reg |= 0x80;
+	HAL_I2C_Mem_Read(handle, HTS221_I2C_ADDRESS, reg, I2C_MEMADD_SIZE_8BIT,
+			bufp, len, 1000);
+	return 0;
+}
+
+void SENSOR_hts221_Read_Data(void){
+	/* Read humidity data */
+	memset(&data_raw_humidity, 0x00, sizeof(int16_t));
+	hts221_humidity_raw_get(&dev_ctx_hts221, &data_raw_humidity);
+	humidity_perc = linear_interpolation(&lin_hum, data_raw_humidity);
+
+	if (humidity_perc < 0) {
+		humidity_perc = 0;
+	}
+	if (humidity_perc > 100) {
+		humidity_perc = 100;
+	}
+	sprintf((char*)tx_buffer, "Humidity [%%]:%3.2f\r\n", humidity_perc);
+	tx_com(tx_buffer, strlen((char const*)tx_buffer));
+
+	/* Read temperature data */
+	memset(&data_raw_temperature, 0x00, sizeof(int16_t));
+	hts221_temperature_raw_get(&dev_ctx_hts221, &data_raw_temperature);
+	temperature_degC = linear_interpolation(&lin_temp, data_raw_temperature);
+	sprintf((char*)tx_buffer, "Temperature [degC]:%6.2f\r\n", temperature_degC);
+	tx_com(tx_buffer, strlen((char const*)tx_buffer));
+}
+
+//SENSOR lps22hh
+void SENSOR_lps22hh_Init(void) {
+	lps22hh_pin_int_route_t int_route;
+
+	/* Initialize mems driver interface */
+	dev_ctx_lps22hh.write_reg = lps22hh_write;
+	dev_ctx_lps22hh.read_reg = lps22hh_read;
+	dev_ctx_lps22hh.handle = & SENSOR_BUS;
+
+	/* Wait sensor boot time */
+	platform_delay(BOOT_TIME);
+	/* Check device ID */
+	whoamI = 0;
+	int result = lps22hh_device_id_get(&dev_ctx_lps22hh, &whoamI);
+
+	if (result == 0) {
+		printf("WHO_AM_I: %d\r\n", whoamI);
+	} else {
+		printf("Error reading WHO_AM_I register\r\n");
+	}
+
+	if (whoamI != LPS22HH_ID)
+		while (1)
+			; /*manage here device not found */
+
+	/* Restore default configuration */
+	lps22hh_reset_set(&dev_ctx_lps22hh, PROPERTY_ENABLE);
+
+	do {
+		lps22hh_reset_get(&dev_ctx_lps22hh, &rst);
+	} while (rst);
+
+	lps22hh_int_notification_set(&dev_ctx_lps22hh, LPS22HH_INT_PULSED);
+
+	lps22hh_pin_int_route_get(&dev_ctx_lps22hh, &int_route);
+	int_route.drdy_pres = PROPERTY_ENABLE;
+	lps22hh_pin_int_route_set(&dev_ctx_lps22hh, &int_route);
+
+	/* Enable Block Data Update */
+	lps22hh_block_data_update_set(&dev_ctx_lps22hh, PROPERTY_ENABLE);
+	/* Set Output Data Rate */
+	lps22hh_data_rate_set(&dev_ctx_lps22hh, LPS22HH_ONE_SHOOT);
+}
+
+int32_t lps22hh_write(void *handle, uint8_t reg, const uint8_t *bufp,
+		uint16_t len) {
+	HAL_I2C_Mem_Write(handle, LPS22HH_I2C_ADD_H, reg, I2C_MEMADD_SIZE_8BIT,
+			(uint8_t*) bufp, len, 1000);
+	return 0;
+}
+
+int32_t lps22hh_read(void *handle, uint8_t reg, uint8_t *bufp,
+		uint16_t len) {
+	HAL_I2C_Mem_Read(handle, LPS22HH_I2C_ADD_H, reg, I2C_MEMADD_SIZE_8BIT, bufp,
+			len, 1000);
+	return 0;
+}
+
+void SENSOR_lps22hh_Read_Data(void){
+	memset(&data_raw_pressure, 0x00, sizeof(uint32_t));
+	lps22hh_pressure_raw_get(&dev_ctx_lps22hh, &data_raw_pressure);
+	pressure_hPa = lps22hh_from_lsb_to_hpa(data_raw_pressure);
+	sprintf((char*)tx_buffer, "Pressure [hPa]:%6.2f\r\n", pressure_hPa);
+	tx_com(tx_buffer, strlen((char const*)tx_buffer));
+}
+
+//SENSOR wind speed
+void SENSOR_WindSpeed_Read_Data(void){
+	uint16_t timer_counter = __HAL_TIM_GET_COUNTER(&htim8);
+	windspeed_kph = calculateWindSpeed(timer_counter);
+
+	printf("Wind Speed [kph]:%f\r\n", windspeed_kph);
+	__HAL_TIM_SET_COUNTER(&htim8, 0);
+}
+
+float calculateWindSpeed(uint16_t switchClosures) {
+	// Conversion factor: 1.492 mph corresponds to 1 switch closure per second
+	float countsPerSecond = (float) switchClosures / 60.0;
+	// Conversion factor: 1 mph = 1.60934 kph
+	float windSpeedKph = countsPerSecond * 1.492 * 1.60934;
+
+	return windSpeedKph;
+}
+
+//SENSOR wind direction
+void SENSOR_WindDir_Read_Data(void){
+	const char *direction = determineDirection(adc_buf[0]);
+	printf("Wind Direction: %s\r\n", direction);
+}
+
+const char* determineDirection(uint16_t adcValue) {
+	// Calculate the sector size based on the number of directions
+	const int numDirections = 16;
+	const int sectorSize = 4096 / numDirections;
+
+	// Calculate the sector index based on the ADC value
+	int sectorIndex = (adcValue + sectorSize / 2) / sectorSize;
+
+	// Define the compass directions
+	const char *compassDirections[] = { "N", "NNE", "NE", "ENE", "E", "ESE",
+			"SE", "SSE", "S", "SSO", "SO", "OSO", "O", "ONO", "NO", "NNO" };
+
+	// Map the sector index to the corresponding direction
+	return compassDirections[sectorIndex % numDirections];
+}
+
+//SENSOR rainfall
+void SENSOR_Rain_Read_Data(void){
+	uint16_t timer_counter = __HAL_TIM_GET_COUNTER(&htim2);
+	rainfall_mm = calculateRainfall(timer_counter);
+
+	printf("Rainfall [mm]:%f\r\n", rainfall_mm);
+	__HAL_TIM_SET_COUNTER(&htim2, 0);
+}
+
+float calculateRainfall(uint16_t switchClosures) {
+	// Conversion factor: 1 switch closure corresponds to 0.011" (0.2794 mm) of rain
+	float rainInches = (float) switchClosures * 0.011;
+	// Conversion factor: 1 inch = 25.4 mm
+	float rainMillimeters = rainInches * 25.4;
+
+	return rainMillimeters;
+}
